requirements.txt
gunicorn==19.0.0
Django==1.10.4
future==0.16.0
line-bot-sdk==1.0.2
requests==2.12.3 
